requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
python-multipart==0.0.20
pillow==12.0.0
openai==2.5.0
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
imagehash==4.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
